added script for text show stamina and health values

+++ Assets/Scripts/Player/Player.cs
@@ -5,7 +5,7 @@ using Input = UnityEngine.Input;
 
 public class Player : CharacterBlueprint
 {
-    public Player(string name, int health, int currentHealth, int damage, float stamina, float currentStamina)
+    public Player(string name, int health, int currentHealth, int damage, int stamina, double currentStamina)
     {
         Name = name;
         Health = health;
